Migrate to new AST printer
